+++ app/src/main/java/com/example/team17zooseeker/ItineraryActivity.java
@@ -7,6 +7,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 
@@ -27,6 +28,10 @@ public class ItineraryActivity extends AppCompatActivity {
     private ArrayList<String> VList;
     private Set<String> VSet;
 
+    private Button clearBtn;
+
+    private static boolean currentlyTesting = false;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -40,8 +45,11 @@ public class ItineraryActivity extends AppCompatActivity {
         database = ZooKeeperDatabase.getSingleton(this);
         stateDao = database.stateDao();
 
-        stateDao.delete(stateDao.get());
-        stateDao.insert(new State("1"));
+        // For Preservation Testing
+        if(!currentlyTesting) {
+            stateDao.delete(stateDao.get());
+            stateDao.insert(new State("1"));
+        }
 
         // For Itinerary Activity
         preferences = getPreferences(MODE_PRIVATE);
@@ -83,6 +91,9 @@ public class ItineraryActivity extends AppCompatActivity {
         //Setting the directions button event
         this.Get_direction = this.findViewById(R.id.get_direction);
         Get_direction.setOnClickListener(this::onGetDirectionClicked);
+
+        clearBtn = findViewById(R.id.clear_btn2);
+        clearBtn.setOnClickListener(this::onClearClicked);
     }
 
     void onGetDirectionClicked (View view){
@@ -93,6 +104,35 @@ public class ItineraryActivity extends AppCompatActivity {
         editor.putStringSet("VList", null);
         editor.apply();
 
+        finish();
+    }
+
+    public void onClearClicked (View view) {
+        Itinerary.deleteItinerary();
+        Itinerary.setItineraryCreated(false);
+        //Setting current index position
+        Directions.resetCurrentIndex();
+
+        stateDao.delete(stateDao.get());
+        stateDao.insert(new State("0"));
+
+        editor.putStringSet("VList", null);
+        editor.putInt("ItinIndex", 0);
+        editor.apply();
+
+        Intent intent = new Intent(this, MainActivity.class);
+        startActivity(intent);
+        finish();
+    }
+
+    public static void setTesting(boolean s) {
+        currentlyTesting = s;
+    }
+
+    public void testDirection(HashSet<String> s) {
+        Intent intent = new Intent(this, DirectionsActivity.class);
+        intent.putExtra("VList", new ArrayList(s));
+        startActivity(intent);
         finish();
     }
 }
